backend/requirements.txt
flask==2.2.3
flask-sqlalchemy==3.0.3
flask-cors==3.0.10
flask-jwt-extended==4.4.4
python-dotenv==1.0.0
gunicorn==20.1.0
psycopg2-binary>=2.9.5
SQLAlchemy==2.0.4
Werkzeug==2.2.3
